+++ core/routes/__init__.py
@@ -1,5 +1,6 @@
 from .callback import callback_namespace
 from .debug import debug_namespace
 from .trial_network import trial_network_namespace
+from .user_no_two_factor import user_no_two_factor_namespace
 from .user import user_namespace
 from .verification_token import verification_token_namespace